Create script.js
I moved the original <script> into a script.js to make it easier to read and this is how I worked on it mostly. I only put it back into the HTML file right before uploading because that's how you originally had it. I have the original script commented out along with the changes I made original before changing the new list items to a paragraph element and a few other changes.